requirements.txt
pandas==2.3.3
plotly==6.6.0
psycopg2-binary==2.9.11
python-dotenv==1.2.2
requests==2.33.0
scipy==1.17.1
SQLAlchemy==2.0.48
streamlit==1.55.0
